+++ berkeley-mobile/Events/EventsView.swift
@@ -17,23 +17,26 @@ struct EventsView: View {
             BMTopBlobView(imageName: "BlobTopRight", xOffset: 50, yOffset: -45, width: 300, height: 150)
             NavigationStack {
                 if eventsViewModel.eventsGroupedByDate.isEmpty {
-                    BMNoEventsView()
+                    BMContentUnavailableView(iconName: "", title: "No Events Available", subtitle: "Please try again.")
                 } else {
                     List(eventsViewModel.eventsGroupedByDate, id: \.0) { dayEvents in
                         EventsDateSectionView(date: dayEvents.0, events: dayEvents.1) { event in
-                            if event.isAllDay == true {
-                                AllDayEventBannerView(event: event)
-                            } else {
-                                EventRowView(event: event)
-                                    .padding(.horizontal)
-                                    .background(
-                                        NavigationLink("") {
-                                            EventDetailView(event: event)
-                                                .presentAlert(alert: $eventsViewModel.alert)
-                                        }
-                                        .opacity(0)
-                                    )
+                            Group {
+                                if event.isAllDay == true {
+                                    AllDayEventBannerView(event: event)
+                                } else {
+                                    EventRowView(event: event)
+                                        .padding(.horizontal)
+
+                                }
                             }
+                            .background(
+                                NavigationLink("") {
+                                    EventDetailView(event: event)
+                                        .presentAlert(alert: $eventsViewModel.alert)
+                                }
+                                .opacity(0)
+                            )
                         }
                         .listRowSeparator(.hidden)
                         .listRowBackground(Color.clear)
@@ -52,34 +55,4 @@ struct EventsView: View {
 
 #Preview {
     EventsView()
-}
-
-struct AllDayEventBannerView: View {
-    @InjectedObservable(\.eventsViewModel) private var eventsViewModel
-
-    var event: BMEventCalendarEntry
-
-    var body: some View {
-        Capsule()
-            .fill(.gray.opacity(0.5))
-            .frame(height: 30)
-            .overlay(
-                HStack(spacing: 10) {
-                    Text("All Day")
-                        .font(Font(BMFont.bold(15)))
-                    Text(event.name)
-                        .font(Font(BMFont.regular(12)))
-                    Spacer()
-                }
-                .padding(.horizontal, 10)
-                .padding(.vertical, 4)
-                .background(
-                    NavigationLink("") {
-                        EventDetailView(event: event)
-                            .presentAlert(alert: $eventsViewModel.alert)
-                    }
-                    .opacity(0)
-                )
-            )
-    }
 }
